quantrautil.py
This file is created to fetch the data from different web sources. If fetching from one of the web sources fails then it tries another web source. The code needs to be stored as quantrautil.py. And save the file in .....\Anaconda2\Lib\site-packages folder or in your current working folder. You have to install the necessary packages and the line no 13 needs to be replaced with your Quandl key for the code to work correctly